+++ config/dev.exs
@@ -16,6 +16,7 @@ config :lightning, Lightning.Repo,
 # watchers to your application. For example, we use it
 # with esbuild to bundle .js and .css sources.
 config :lightning, LightningWeb.Endpoint,
+  http: [ip: {127, 0, 0, 1}, port: 4000],
   check_origin: false,
   code_reloader: true,
   debug_errors: true,