+++ lib/lightning/auditing.ex
@@ -5,10 +5,12 @@ defmodule Lightning.Auditing do
 
   import Ecto.Query
   alias Lightning.Repo
+  alias Lightning.Auditing.Model, as: Audit
+  alias Lightning.Accounts.User
 
   def list_all(params \\ %{}) do
-    from(a in Lightning.Credentials.Audit.base_query(),
-      left_join: u in Lightning.Accounts.User,
+    from(a in Audit,
+      left_join: u in User,
       on: [id: a.actor_id],
       select_merge: %{actor: u},
       order_by: [desc: a.inserted_at]