+++ lib/lightning/invocation.ex
@@ -14,6 +14,9 @@ defmodule Lightning.Invocation do
   alias Lightning.Projects.Project
 
   @workorders_search_timeout 30_000
+  @workorders_count_limit 50
+
+  def get_workorders_count_limit, do: @workorders_count_limit
 
   @doc """
   Returns the list of dataclips.
@@ -280,22 +283,6 @@ defmodule Lightning.Invocation do
   end
 
   @doc """
-  Deletes a run.
-
-  ## Examples
-
-      iex> delete_run(run)
-      {:ok, %Run{}}
-
-      iex> delete_run(run)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_run(%Run{} = run) do
-    Repo.delete(run)
-  end
-
-  @doc """
   Returns an `%Ecto.Changeset{}` for tracking run changes.
 
   ## Examples
@@ -328,16 +315,23 @@ defmodule Lightning.Invocation do
 
   def search_workorders(
         %Project{} = project,
-        %SearchParams{} = search_params,
+        %SearchParams{search_term: search_term} = search_params,
         params \\ %{}
       ) do
     params =
       update_in(
         params,
         [:options],
-        fn
-          nil -> [timeout: @workorders_search_timeout]
-          options -> options
+        fn options ->
+          [timeout: @workorders_search_timeout]
+          |> Keyword.merge(options || [])
+          |> then(fn options ->
+            if search_term do
+              Keyword.put(options, :limit, @workorders_count_limit)
+            else
+              options
+            end
+          end)
         end
       )