+++ test/lightning_web/live/work_order_live_test.exs
@@ -238,29 +238,29 @@ defmodule LightningWeb.RunWorkOrderTest do
 
       assert html =~ "Status"
 
-      assert view
-             |> element("input#run-filter-form_success[checked]")
-             |> has_element?()
+      status_filters = [
+        "success",
+        "failed",
+        "running",
+        "crashed",
+        "pending",
+        "killed",
+        "failed"
+      ]
 
-      assert view
-             |> element("input#run-filter-form_failed[checked]")
-             |> has_element?()
+      assert status_filters
+             |> Enum.all?(fn f ->
+               view
+               |> element("input#workorder-filter-form_#{f}")
+               |> has_element?()
+             end)
 
-      assert view
-             |> element("input#run-filter-form_running[checked]")
-             |> has_element?()
-
-      assert view
-             |> element("input#run-filter-form_crashed[checked]")
-             |> has_element?()
-
-      assert view
-             |> element("input#run-filter-form_pending[checked]")
-             |> has_element?()
-
-      assert view
-             |> element("input#run-filter-form_killed[checked]")
-             |> has_element?()
+      assert status_filters
+             |> Enum.any?(fn f ->
+               view
+               |> element("input#workorder-filter-form_#{f}[checked]")
+               |> has_element?()
+             end) == false
 
       assert view
              |> element("input#run-search-form_search_term")
@@ -306,7 +306,7 @@ defmodule LightningWeb.RunWorkOrderTest do
              |> has_element?()
     end
 
-    test "Workorder with failed status shows when option checked", %{
+    test "Work Order with failed status shows when option checked", %{
       conn: conn,
       project: project
     } do
@@ -351,7 +351,7 @@ defmodule LightningWeb.RunWorkOrderTest do
 
       # uncheck :failure
       view
-      |> form("#run-filter-form", filters: %{"failed" => "false"})
+      |> form("#workorder-filter-form", filters: %{"failed" => "false"})
       |> render_submit()
 
       refute view
@@ -363,7 +363,7 @@ defmodule LightningWeb.RunWorkOrderTest do
       # recheck failure
 
       view
-      |> form("#run-filter-form", filters: %{"failed" => "true"})
+      |> form("#workorder-filter-form", filters: %{"failed" => "true"})
       |> render_submit()
 
       div =
@@ -1090,7 +1090,9 @@ defmodule LightningWeb.RunWorkOrderTest do
       {:ok, view, _html} =
         live_async(
           conn,
-          Routes.project_run_index_path(conn, :index, project.id)
+          Routes.project_run_index_path(conn, :index, project.id,
+            filters: %{workflow_id: workflow_1.id}
+          )
         )
 
       work_order_1 =
@@ -1142,35 +1144,23 @@ defmodule LightningWeb.RunWorkOrderTest do
 
       Events.subscribe(project.id)
 
+      # send a workorder that matches the current filter criteria
       Events.work_order_created(project.id, work_order_1)
       %{id: wo_id} = work_order_1
       assert_received %Events.WorkOrderCreated{work_order: %{id: ^wo_id}}
 
-      assert_patch(
-        view,
-        Routes.project_run_index_path(conn, :index, project.id,
-          filters: Map.put(filters, "workorder_id", wo_id)
-        )
-      )
-
       # Awaits for async changes and forces re-render
       render_async(view)
       render(view)
+
       assert has_element?(view, "#workorder-#{work_order_1.id}")
 
-      # repeat same test for another workorder
+      # repeat same test for another workorder and show that it does not appear
       Events.work_order_created(project.id, work_order_2)
       %{id: wo_id} = work_order_2
       assert_received %Events.WorkOrderCreated{work_order: %{id: ^wo_id}}
 
-      assert_patch(
-        view,
-        Routes.project_run_index_path(conn, :index, project.id,
-          filters: Map.put(filters, "workorder_id", wo_id)
-        )
-      )
-
-      assert has_element?(view, "#workorder-#{work_order_2.id}")
+      refute has_element?(view, "#workorder-#{work_order_2.id}")
     end
   end
 
@@ -1217,6 +1207,12 @@ defmodule LightningWeb.RunWorkOrderTest do
       workflow = insert(:workflow, project: project)
       trigger = insert(:trigger, type: :webhook, workflow: workflow)
       job_a = insert(:job, workflow: workflow)
+
+      insert(:edge,
+        workflow: workflow,
+        source_trigger: trigger,
+        target_job: job_a
+      )
 
       dataclip = insert(:dataclip)
 
@@ -1412,6 +1408,7 @@ defmodule LightningWeb.RunWorkOrderTest do
       render_change(view, "toggle_all_selections", %{all_selections: true})
       result = render_click(view, "bulk-rerun", %{type: "all"})
       {:ok, view, html} = follow_redirect(result, conn)
+      render_async(view)
 
       assert html =~ "New attempts enqueued for 5 workorders"
 
@@ -1897,11 +1894,11 @@ defmodule LightningWeb.RunWorkOrderTest do
       render_async(view)
 
       refute html =~
-               "Find all runs that include this step, and rerun from there"
+               "Find all runs that include this step and rerun from there"
 
       assert render_change(view, "toggle_all_selections", %{
                all_selections: true
-             }) =~ "Find all runs that include this step, and rerun from there"
+             }) =~ "Find all runs that include this step and rerun from there"
 
       view
       |> form("#select-job-for-rerun-form")
@@ -1913,6 +1910,8 @@ defmodule LightningWeb.RunWorkOrderTest do
 
       assert html =~
                "New attempts enqueued for 2 workorders"
+
+      render_async(view)
 
       view
       |> form("#selection-form-#{work_order_1.id}")
@@ -1927,7 +1926,8 @@ defmodule LightningWeb.RunWorkOrderTest do
 
       {:ok, _view, html} = follow_redirect(result, conn)
 
-      assert html =~ "New attempt enqueued for 1 workorder"
+      # this is zero because the previous retried attempt has no run
+      assert html =~ "New attempt enqueued for 0 workorder"
     end
 
     test "jobs on the modal are updated every time the selected workflow is changed",